Refactor/Extract OpenApi bootstrapping from Program.cs
