+++ src/Api/Program.cs
@@ -11,6 +11,7 @@ using Passwordless.Api.Extensions;
 using Passwordless.Api.HealthChecks;
 using Passwordless.Api.Helpers;
 using Passwordless.Api.Middleware;
+using Passwordless.Api.OpenApi;
 using Passwordless.Api.OpenApi.Filters;
 using Passwordless.Api.RateLimiting;
 using Passwordless.Api.Reporting.Background;
@@ -28,36 +29,7 @@ using Passwordless.Service.Storage.Ef;
 WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
 builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen(swagger =>
-{
-    swagger.IncludeXmlComments(Path.Combine(AppContext.BaseDirectory, "Passwordless.Api.xml"), true);
-    swagger.DocInclusionPredicate((docName, apiDesc) =>
-    {
-        var policy = (AuthorizationPolicy?)apiDesc.ActionDescriptor.EndpointMetadata.FirstOrDefault(x => x.GetType() == typeof(AuthorizationPolicy));
-        if (policy == null)
-        {
-            return false;
-        }
-        return !policy.AuthenticationSchemes.Contains(Constants.ManagementKeyAuthenticationScheme);
-    });
-    swagger.OperationFilter<AuthorizationOperationFilter>();
-    swagger.OperationFilter<ExtendedStatusDescriptionsOperationFilter>();
-    swagger.OperationFilter<ExternalDocsOperationFilter>();
-    swagger.SupportNonNullableReferenceTypes();
-    swagger.SwaggerDoc("v4", new OpenApiInfo
-    {
-        Version = "v4",
-        Title = "Passwordless.dev API",
-        TermsOfService = new Uri("https://bitwarden.com/terms/"),
-        Contact = new OpenApiContact
-        {
-            Email = "support@passwordless.dev",
-            Name = "Support",
-            Url = new Uri("https://bitwarden.com/contact/")
-        }
-    });
-    swagger.SwaggerGeneratorOptions.IgnoreObsoleteActions = true;
-});
+builder.Services.AddOpenApi();
 
 if (builder.Configuration.IsSelfHosted())
 {
@@ -147,21 +119,7 @@ else
             "Hey, this place is for computers. Check out our human documentation instead: https://docs.passwordless.dev");
 }
 
-app.UseSwagger(c => c.PreSerializeFilters.Add((swaggerDoc, httpReq) =>
-{
-    httpReq.HttpContext.Response.Headers.Append("Access-Control-Allow-Origin", "*");
-}));
-
-app.UseSwaggerUI(c =>
-{
-    c.SwaggerEndpoint("/swagger/v4/swagger.json", "v4");
-    c.ConfigObject.ShowExtensions = true;
-    c.ConfigObject.ShowCommonExtensions = true;
-    c.DefaultModelsExpandDepth(-1);
-    c.IndexStream = () => typeof(Program).Assembly.GetManifestResourceStream("Passwordless.Api.OpenApi.swagger.html");
-    c.InjectStylesheet("/openapi.css");
-    c.SupportedSubmitMethods();
-});
+app.UseOpenApi();
 
 if (builder.Configuration.IsSelfHosted())
 {
